cookbook/pocoflow-tool-pdf-vision/requirements.txt
pocoflow>=0.2.0
click>=8.0
openai>=1.0.0
PyMuPDF>=1.22.0
Pillow>=10.0.0
